pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "clayarg"
version = "0.3.0"
description = "Full-local macOS photogrammetry pipeline"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "numpy>=1.26",
    "Pillow>=10.0",
    "opencv-python-headless>=4.8",
    "pymeshlab>=2023.12",
    "pyobjc-framework-Vision>=10.0",
    "pyobjc-framework-Cocoa>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[project.scripts]
clayarg = "clayarg.cli:main"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
